Extract `StringHook` from `FrameworkHook`
